Fix macOS build, UI text visibility, audio permissions, and warnings
- Add NSMicrophoneUsageDescription to Info-App.plist so macOS shows the
  audio permission prompt (without it, audio access is silently denied,
  blocking both input and output)
- Add hidapi_darwin.h (missing header needed for HIDAPI to compile on macOS)
- Add HidApiPlatform.c to Xcode project and set hidapi header search paths
  in project.pbxproj
- Fix invisible text on Default dark theme at startup: call colorSchemeChanged()
  after addListener() in MainComponent constructor so TextEditor cached colors
  are refreshed at launch, not only on theme switch
- Fix first-party build warnings: unused variables in WfsStandardSlider,
  WfsWidthExpansionSlider, InputsTabPages; implicit size_t conversion in
  AudioInterfaceWindow; comment-nesting warnings in OSCMessageRouter

+++ Source/gui/AudioInterfaceWindow.cpp
@@ -483,7 +483,7 @@ void DeviceSettingsPanel::sampleRateChanged()
     {
         juce::AudioDeviceManager::AudioDeviceSetup setup;
         deviceManager.getAudioDeviceSetup(setup);
-        setup.sampleRate = rates[static_cast<size_t>(index)];
+        setup.sampleRate = rates[index];
 
         auto error = deviceManager.setAudioDeviceSetup(setup, true);
         if (!error.isEmpty())
@@ -513,7 +513,7 @@ void DeviceSettingsPanel::bufferSizeChanged()
     {
         juce::AudioDeviceManager::AudioDeviceSetup setup;
         deviceManager.getAudioDeviceSetup(setup);
-        setup.bufferSize = sizes[static_cast<size_t>(index)];
+        setup.bufferSize = sizes[index];
 
         auto error = deviceManager.setAudioDeviceSetup(setup, true);
         if (!error.isEmpty())
